+++ packages/protocol/src/dcl/index.ts
@@ -8,6 +8,7 @@ export * from "./DclCertificateService.js";
 export * from "./DclClient.js";
 export * from "./DclOtaUpdateService.js";
 export * from "./DclRestApiTypes.js";
+export * from "./DclVendorInfoService.js";
 export * from "./OtaImageHeader.js";
 export * from "./OtaImageReader.js";
 export * from "./OtaImageWriter.js";